重新规划

+++ reactor-server/src/main/java/com/doopp/agar/handle/AgarHandle.java
@@ -54,8 +54,8 @@ public class AgarHandle extends AbstractWebSocketServerHandle {
         User user = getChannelUser(channel);
         if (user!=null && !Strings.isNullOrEmpty(frame.text()) && frame.text().contains(" ")) {
             String[] xy = frame.text().split(" ");
-            user.setX(Float.parseFloat(xy[0]));
-            user.setY(Float.parseFloat(xy[1]));
+            user.setX(Integer.parseInt(xy[0]));
+            user.setY(Integer.parseInt(xy[1]));
             players.put(user.getId(), user);
         }
         return Mono.empty();
